refactor(registry): split preview and appearance rules
