+++ src/preview/code/registry/data/tooling.rs
@@ -1,0 +1,93 @@
+use super::super::types::{RegistryEntry, entry, language};
+use crate::file_info::CodeBackend;
+
+pub(super) const LANGUAGES: &[RegistryEntry] = &[
+    entry(
+        language("sql", "SQL", CodeBackend::Syntect, None),
+        &["sql"],
+        &[],
+        &[],
+        &["sql"],
+        &["sql"],
+    ),
+    entry(
+        language("diff", "Diff", CodeBackend::Syntect, None),
+        &["diff", "patch"],
+        &[],
+        &[],
+        &["diff", "patch"],
+        &["diff", "patch"],
+    ),
+    entry(
+        language("dockerfile", "Dockerfile", CodeBackend::Syntect, None),
+        &[],
+        &["dockerfile", "containerfile"],
+        &[],
+        &["dockerfile"],
+        &["dockerfile", "docker"],
+    ),
+    entry(
+        language("hcl", "HCL", CodeBackend::Syntect, None),
+        &["hcl"],
+        &[".terraform.lock.hcl"],
+        &[],
+        &["hcl"],
+        &["hcl"],
+    ),
+    entry(
+        language("terraform", "Terraform", CodeBackend::Syntect, None),
+        &["tf", "tfvars", "tfbackend"],
+        &["terraform.rc", ".terraformrc"],
+        &[],
+        &["terraform", "tf", "tfvars"],
+        &["terraform", "tf", "tfvars"],
+    ),
+    entry(
+        language("groovy", "Groovy", CodeBackend::Syntect, None),
+        &["groovy", "gvy", "gradle"],
+        &["build.gradle", "settings.gradle", "init.gradle"],
+        &["groovy"],
+        &["groovy", "gradle"],
+        &["groovy", "gradle"],
+    ),
+    entry(
+        language("scala", "Scala", CodeBackend::Syntect, None),
+        &["scala", "sbt"],
+        &["build.sbt"],
+        &["scala"],
+        &["scala", "sbt"],
+        &["scala", "sbt"],
+    ),
+    entry(
+        language("just", "Just", CodeBackend::Syntect, None),
+        &[],
+        &["justfile", ".justfile"],
+        &[],
+        &["just"],
+        &["just"],
+    ),
+    entry(
+        language("make", "Makefile", CodeBackend::Syntect, None),
+        &["mk", "mak"],
+        &["makefile", "gnumakefile", "bsdmakefile"],
+        &[],
+        &["make", "makefile"],
+        &["make", "makefile"],
+    ),
+    entry(
+        language("nix", "Nix", CodeBackend::Syntect, None),
+        &["nix"],
+        &[],
+        &[],
+        &["nix"],
+        &["nix"],
+    ),
+    entry(
+        language("cmake", "CMake", CodeBackend::Syntect, None),
+        &["cmake"],
+        &["cmakelists.txt"],
+        &[],
+        &["cmake"],
+        &["cmake"],
+    ),
+];
