+++ src/preview/code/registry/data/web.rs
@@ -1,0 +1,85 @@
+use super::super::types::{RegistryEntry, entry, language};
+use crate::file_info::CodeBackend;
+
+pub(super) const LANGUAGES: &[RegistryEntry] = &[
+    entry(
+        language("html", "HTML", CodeBackend::Syntect, None),
+        &["html", "htm", "xhtml"],
+        &[],
+        &[],
+        &["html"],
+        &["html"],
+    ),
+    entry(
+        language("xml", "XML", CodeBackend::Syntect, None),
+        &["xml", "xsd", "xsl", "xslt", "svg"],
+        &[],
+        &[],
+        &["xml", "svg", "markup"],
+        &["xml", "xhtml", "svg", "markup"],
+    ),
+    entry(
+        language("css", "CSS", CodeBackend::Syntect, None),
+        &["css"],
+        &[],
+        &[],
+        &["css"],
+        &["css"],
+    ),
+    entry(
+        language("scss", "SCSS", CodeBackend::Syntect, None),
+        &["scss"],
+        &[],
+        &[],
+        &["scss"],
+        &["scss"],
+    ),
+    entry(
+        language("sass", "Sass", CodeBackend::Syntect, None),
+        &["sass"],
+        &[],
+        &[],
+        &["sass"],
+        &["sass"],
+    ),
+    entry(
+        language("less", "Less", CodeBackend::Syntect, None),
+        &["less"],
+        &[],
+        &[],
+        &["less"],
+        &["less"],
+    ),
+    entry(
+        language("javascript", "JavaScript", CodeBackend::Syntect, None),
+        &["js", "mjs", "cjs"],
+        &[],
+        &[],
+        &["javascript"],
+        &["js", "javascript"],
+    ),
+    entry(
+        language("jsx", "JSX", CodeBackend::Syntect, None),
+        &["jsx"],
+        &[],
+        &[],
+        &["jsx"],
+        &["jsx"],
+    ),
+    entry(
+        language("typescript", "TypeScript", CodeBackend::Syntect, None),
+        &["ts", "mts", "cts"],
+        &[],
+        &[],
+        &["typescript"],
+        &["ts", "typescript"],
+    ),
+    entry(
+        language("tsx", "TSX", CodeBackend::Syntect, None),
+        &["tsx"],
+        &[],
+        &[],
+        &["tsx"],
+        &["tsx"],
+    ),
+];
